src/main/java/com/natsumes/leetcode/topic/ArraySeries.java
package com.natsumes.leetcode.topic;

/**
 * @author hetengjiao
 */
public class ArraySeries {
    /**
     * 给定一个数组 nums，编写一个函数将所有 0 移动到数组的末尾，同时保持非零元素的相对顺序。
     *
     * 示例:
     *
     * 输入: [0,1,0,3,12]
     * 输出: [1,3,12,0,0]
     * 说明:
     *
     * 必须在原数组上操作，不能拷贝额外的数组。
     * 尽量减少操作次数。
     *
     * 作者：力扣 (LeetCode)
     * 链接：https://leetcode-cn.com/leetbook/read/all-about-array/x9rh8e/
     * 来源：力扣（LeetCode）
     * 著作权归作者所有。商业转载请联系作者获得授权，非商业转载请注明出处。
     */
    public void moveZeroes(int[] nums) {

        int i = 0, j = 1;
        while (i < nums.length && j < nums.length) {
            if (nums[i] != 0) {
                i++;
            }
            if (nums[i] == 0 && nums[j] != 0) {
                nums[i] = nums[j];
                nums[j] = 0;
                i++;
            }
            j++;
        }
    }


}
